Add `pipeline::activity` operator
